backend/requirements.txt
# RepoWhisper Backend Dependencies
# Python 3.12+

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# JWT Auth
python-jose[cryptography]==3.3.0

# AI/ML - Transcription & Embeddings
faster-whisper==1.0.3
sentence-transformers==3.1.1
torch>=2.0.0  # Flexible version for compatibility

# Vector Database
lancedb==0.13.0
pyarrow==17.0.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Security & Rate Limiting
slowapi==0.1.9
structlog==24.1.0

# AI/LLM for Boss Mode (Groq uses OpenAI-compatible API)
openai==1.12.0
pillow>=11.0.0

